include/types.h: move errorcode32_t to separate header
