+++ src/crimson/CMakeLists.txt
@@ -106,6 +106,7 @@ add_library(crimson-common STATIC
   ${PROJECT_SOURCE_DIR}/src/include/filepath.cc
   ${PROJECT_SOURCE_DIR}/src/include/object.cc
   ${PROJECT_SOURCE_DIR}/src/include/types.cc
+  ${PROJECT_SOURCE_DIR}/src/include/errorcode32.cc
   ${PROJECT_SOURCE_DIR}/src/include/utime.cc
   ${PROJECT_SOURCE_DIR}/src/include/uuid.cc
   ${PROJECT_SOURCE_DIR}/src/librbd/Features.cc
