+++ src/include/types.cc
@@ -124,14 +124,3 @@ std::list<shard_id_t> shard_id_t::generate_test_instances() {
   ls.push_back(shard_id_t(2));
   return ls;
 }
-
-void errorcode32_t::dump(ceph::Formatter *f) const {
-  f->dump_int("code", code);
-}
-
-std::list<errorcode32_t> errorcode32_t::generate_test_instances() {
-  std::list<errorcode32_t> ls;
-  ls.push_back(errorcode32_t(1));
-  ls.push_back(errorcode32_t(2));
-  return ls;
-}
