+++ src/messages/MClientReply.h
@@ -17,7 +17,7 @@
 #ifndef CEPH_MCLIENTREPLY_H
 #define CEPH_MCLIENTREPLY_H
 
-#include "include/types.h"
+#include "include/errorcode32.h"
 #include "include/fs_types.h"
 #include "include/mempool.h"
 #include "MClientRequest.h"
