+++ src/messages/MMonCommandAck.h
@@ -16,6 +16,7 @@
 #ifndef CEPH_MMONCOMMANDACK_H
 #define CEPH_MMONCOMMANDACK_H
 
+#include "include/errorcode32.h"
 #include "messages/PaxosServiceMessage.h"
 
 #include <sstream>
