+++ src/messages/MWatchNotify.h
@@ -19,6 +19,7 @@
 
 #include <ostream>
 
+#include "include/errorcode32.h"
 #include "msg/Message.h"
