+++ src/osd/osd_types.h
@@ -40,7 +40,7 @@
 #include "msg/msg_types.h"
 #include "include/common_fwd.h" // for CephContext
 #include "include/compat.h"
-#include "include/types.h"
+#include "include/errorcode32.h"
 #include "include/utime.h"
 #include "include/CompatSet.h"
 #include "common/bitset_set.h"
